Finished Chat Page
1. The user can send any messages that will be saved to the database
(and displayed on the screen)
2. The doctor's messages are now retrieved from the database

ISSUES: As the openMRS Rest API doesn't provide an option to get all the
data from the obs table depending on the concept uuid (it also needs the
person's uuid) I have to get the doctor's and patient's messages
separately.
Also, since it doesn't return the date (the JSON only contains the
message uuid and value) I can't order the messages chronologically.
Therefore, I split the screen in 2 parts and fetch the data separately
(top part-patient's messages; bottom part-doctor's messages)